+++ packages/blocks/src/root-block/edgeless/clipboard/clipboard.ts
@@ -1,8 +1,4 @@
-import type {
-  Connection,
-  GfxCompatibleProps,
-  ReferenceParams,
-} from '@blocksuite/affine-model';
+import type { Connection, ReferenceParams } from '@blocksuite/affine-model';
 import type {
   BlockStdScope,
   EditorHost,
@@ -33,6 +29,7 @@ import {
 } from '@blocksuite/affine-shared/utils';
 import {
   compareLayer,
+  type GfxCompatibleProps,
   type SerializedElement,
   SortOrder,
 } from '@blocksuite/block-std/gfx';
@@ -1088,12 +1085,14 @@ export class EdgelessClipboardController extends PageClipboard {
       });
     } else {
       for (let index = 0; index < content.length; index++) {
-        await this.onBlockSnapshotPaste(
-          content[index],
-          this.doc,
-          noteId,
-          index
-        );
+        const blockSnapshot = content[index];
+        if (blockSnapshot.flavour === 'affine:note') {
+          for (const child of blockSnapshot.children) {
+            await this.onBlockSnapshotPaste(child, this.doc, noteId);
+          }
+          continue;
+        }
+        await this.onBlockSnapshotPaste(content[index], this.doc, noteId);
       }
     }